Move GameParseException to gameparser package
