+++ game-core/src/main/java/games/strategy/engine/data/gameparser/GameParser.java
@@ -8,7 +8,6 @@ import games.strategy.engine.data.Attachable;
 import games.strategy.engine.data.EngineVersionException;
 import games.strategy.engine.data.GameData;
 import games.strategy.engine.data.GameMap;
-import games.strategy.engine.data.GameParseException;
 import games.strategy.engine.data.GamePlayer;
 import games.strategy.engine.data.GameStep;
 import games.strategy.engine.data.IAttachment;
