+++ game-core/src/main/java/games/strategy/engine/data/gameparser/XmlReader.java
@@ -1,6 +1,5 @@
 package games.strategy.engine.data.gameparser;
 
-import games.strategy.engine.data.GameParseException;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.URL;
